fix: chakra props; rename master-list component
