+++ static_report/src/components/ComparisonReport/CRTabProfilingDetails.tsx
@@ -4,7 +4,7 @@ import { SaferTableSchema, zReport, ZTableSchema } from '../../types';
 import { getDataChart } from '../../utils/charts';
 import { transformAsNestedBaseTargetRecord } from '../../utils/transformers';
 import { ColumnCardDataVisualContainer } from '../shared/ColumnCard/ColumnCardDataVisualContainer';
-import { CRTableColumnDetailsRow } from './CRTableColumnDetailsRow';
+import { CRColumnDetailsRow } from './CRColumnDetailsRow';
 
 type CRTabProfilingDetailsProps = {
   baseTable?: SaferTableSchema;
@@ -52,7 +52,7 @@ function CRProfilingColumn({ name, base, target }: CRProfilingColumnProps) {
   return (
     <Flex key={name} direction="column">
       <Grid my={8} templateColumns="1fr 2fr" gap={12} overflowX={'hidden'}>
-        <CRTableColumnDetailsRow baseColumn={base} targetColumn={target} />
+        <CRColumnDetailsRow baseColumn={base} targetColumn={target} />
 
         <Flex my={4} alignItems={'center'}>
           <ColumnCardDataVisualContainer height={350}>
