+++ static_report/src/components/shared/ColumnDetailMasterList/index.tsx
